+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.2 — 2026-03-17
 
 - HighCharts modules
+- Field editor UI improvements
 
 ## 1.0.0 — 2026-03-05